feat: added constants to simple and variable product
