+++ src/screens/Product/constants.js
@@ -1,0 +1,11 @@
+const PRODUCT_TYPES = {
+    SIMPLE: "simple",
+    VARIABLE: "variable",
+}
+
+const FORM_FIELDS = {
+    VARIATIONS: "variations",
+    QUANTITY: "quantity",
+}
+
+export { PRODUCT_TYPES, FORM_FIELDS }
